Moved HandData to seperate script
