+++ .idea/.idea.Sushi-2-Go/.idea/dictionaries/wiley.xml
@@ -1,0 +1,3 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="wiley" />
+</component>
