+++ Assets/GameData/HandData/LeftHandData.asset
